(WATCH-20) Create an alert notification on watch
Created a button on the mobile app where when pressed it will pop up a notification on the watch with 3 pages, the initial notification, an extra information page and a page with the image of an EKG graph.
It also creates an extra "Block App" action button at the very last page of the notification. From what I've read this is built into the notifications.