(WATCH-24) Make notification wake up and vibrate
The watch seems to automatically wake up when the vibration occurs. We are able to create our own pattern for the vibration by using the function setVibrate in NotificationCompat.Builder, in the form of {Start delay, Vibration length, Sleep time, Vibration length}, currently it is set to start instantly, vibrate for 3 seconds, sleep for 1 second then vibrate for another 3 seconds.